[upload] postit modification

+++ Namukkun/src/main/java/com/pard/namukkun/postit/controller/PostItController.java
@@ -11,6 +11,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Objects;
+
 @RestController
 @RequestMapping("post/postit")
 @RequiredArgsConstructor
@@ -21,13 +23,21 @@ public class PostItController {
     // 덧글 선택하여 포스트잇으로 만들기
     @GetMapping("/create")
     @Operation(summary = "포스트잇 생성", description = "덧글의 내용을 포스트잇으로 생성합니다.")
-    public ResponseEntity<?> selectCommentToPostIt(PostItCreateDTO dto) {
-        // TODO post 에서 count 증가 해야함
-        postItService.createPostIt(dto);
-        return new ResponseEntity<>(HttpStatus.CREATED);
+    public ResponseEntity<?> selectCommentToPostIt(
+            @RequestParam("userid") Long userid,
+            PostItCreateDTO dto
+    ) {
+        // 권한 없음
+        if (!postItService.getWriterIdByPostIdIt(dto.getPostId()).equals(userid))
+            return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
 
-        // 글쓴이 아니면 권한 없음 return 해야함
-//        return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+        // 최대갯수 확인
+        if (10 <= postItService.getPostPostItCounter(dto.getPostId()))
+            return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE); // 생성 안됨
+
+        // 생성
+        postItService.createPostIt(dto);
+        return new ResponseEntity<>(HttpStatus.CREATED); // 생성됨
     }
 
 
@@ -35,27 +45,22 @@ public class PostItController {
 
 
     @PatchMapping("/sectionmove")
+    @Operation(summary = "포스트잇 섹션 이동", description = "덧글의 섹션을 이동합니다." + "left, right 아니면 값을 받지 않습니다")
     public ResponseEntity<?> moveSectionPostIt(
             @RequestParam("userId") Long userId, // debug
-//            @RequestParam("postId") Long postId, // 글 작성자인지 확인
             @RequestParam("postItId") Long postItId,
             @RequestParam("section") String section) {
         // 로그인 안됨
         if (userId == null) return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 
-        // 권한 없음
-//        if (!userId.equals(postService.getWriterUserId(postItService.getPostId(postidId))))
-//            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
-//
-//        // 보안 대처
-//        if(!postId.equals(postItService.getPostId(postitid))){
-//
-//        }
+        if (!userId.equals(postItService.getWriterIdByPostIdIt(postItId)))
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 
+        // 잘못된 입력
+        if (!(section.equals("left") || section.equals("right")))
+            return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
 
-
-//        postItService.moveSection(postItId, section);
-
+        postItService.moveSection(postItId, section);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
@@ -63,21 +68,31 @@ public class PostItController {
     // 포스트잇 이동
     @PostMapping("/move")
     @Operation(summary = "포스트잇 이동", description = "포스트잇의 위치를 이동시킵니다")
-    public ResponseEntity<?> movePostIt(@RequestBody() PostItMoveDTO dto) {
+    public ResponseEntity<?> movePostIt(
+            @RequestParam("userid") Long userId,
+            @RequestBody() PostItMoveDTO dto
+    ) {
+        // 권한 없음
+        if (!postItService.getWriterIdByPostIdIt(dto.getPostId()).equals(userId))
+            return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+
         postItService.movePostIt(dto);
         return new ResponseEntity<>(HttpStatus.OK);
-        // 글쓴이 아니면 권한 없음 return 해야함
-//        return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
     }
 
     // 포스트잇 삭제
     @DeleteMapping("/delete")
     @Operation(summary = "포스트잇 제거", description = "포스트잇을 지웁니다")
-    public ResponseEntity<?> deletePostIt(@RequestParam("id") Long postItId) {
+    public ResponseEntity<?> deletePostIt(
+            @RequestParam("userid") Long userId,
+            @RequestParam("id") Long postItId
+    ) {
+        // 권한 없음
+        if (!Objects.equals(userId, postItService.getWriterIdByPostIdIt(postItId)))
+            return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+
         // TODO post count --
         postItService.deletePostIt(postItId);
         return new ResponseEntity<>(HttpStatus.OK);
-        // 글쓴이 아니면 권한 없음 return 해야함
-//        return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
     }
 }
